feat: add merkelisation
