+++ client/rpc/src/lib.rs
@@ -1,12 +1,4 @@
-use file_manager::traits::FileStorage;
-use forest_manager::traits::ForestStorage;
-
 use jsonrpsee::core::async_trait;
-use sp_api::ProvideRuntimeApi;
-use sp_runtime::AccountId32;
-use storage_hub_infra::constants::FILE_CHUNK_SIZE;
-use storage_hub_infra::types::Metadata;
-
 use jsonrpsee::core::RpcResult;
 use jsonrpsee::proc_macros::rpc;
 use jsonrpsee::types::error::ErrorObjectOwned as JsonRpseeError;
@@ -14,12 +6,32 @@ use jsonrpsee::types::error::INTERNAL_ERROR_CODE;
 use jsonrpsee::types::error::INTERNAL_ERROR_MSG;
 use jsonrpsee::types::ErrorObjectOwned;
 
+use sc_transaction_pool_api::TransactionPool;
+use sc_transaction_pool_api::TransactionSource;
+use sc_client_api::backend::{Backend, StorageProvider};
+use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_core::H256;
+use sp_block_builder::BlockBuilder as BlockBuilderApi;
 use sp_runtime::traits::Block as BlockT;
+use sp_runtime::BoundedVec;
+use sp_runtime::AccountId32;
+use sp_trie::MemoryDB;
+use sp_trie::TrieDBMutBuilder;
+use sp_runtime::generic::UncheckedExtrinsic;
+use storage_hub_infra::constants::FILE_CHUNK_SIZE;
+use storage_hub_infra::types::Metadata;
+use storage_hub_runtime::AccountId;
+use storage_hub_runtime::RuntimeCall;
+
+use file_manager::traits::FileStorage;
+use forest_manager::traits::ForestStorage;
 
 use log::debug;
 use log::error;
+use serde::de::DeserializeOwned;
+use storage_hub_runtime::Signature;
+use storage_hub_runtime::SignedExtra;
 use std::fmt::Debug;
 use std::fs::File;
 use std::io::Read;
@@ -29,45 +41,35 @@ use tokio::sync::RwLock;
 
 const LOG_TARGET: &str = "file-system-rpc";
 
+
+// CHANGE NAME
 #[rpc(server, namespace = "filesystem")]
 #[async_trait]
-pub trait FileSystemApi<BlockHash> {
+pub trait FileSystemApi {
     #[method(name = "sendFile")]
-    async fn send_file(&self, at: Option<BlockHash>, file_path: String, location: String, owner: AccountId32) -> RpcResult<()>;
+    async fn upload_file(&self, file_path: String, location: String, owner: AccountId32) -> RpcResult<()>;
 }
 
-pub struct FileSystemRpc<C, B, FL, FS> {
-    client: Arc<C>,
+pub struct FileSystemRpc<FL> {
     file_storage: Arc<RwLock<FL>>,
-    forest_storage: Arc<RwLock<FS>>,
-    _marker: std::marker::PhantomData<B>,
 }
 
-impl<C, B, FL, FS> FileSystemRpc<C, B, FL, FS> {
+impl<FL> FileSystemRpc<FL> {
     pub fn new(
-        client: Arc<C>,
         file_storage: Arc<RwLock<FL>>,
-        forest_storage: Arc<RwLock<FS>>,
     ) -> Self {
         Self {
-            client,
             file_storage,
-            forest_storage,
-            _marker: Default::default(),
         }
     }
 }
 
 #[async_trait]
-impl<C, B, FL, FS> FileSystemApiServer<<B as BlockT>::Hash> for FileSystemRpc<C, B, FL, FS>
+impl<FL> FileSystemApiServer for FileSystemRpc<FL>
 where
-    B: BlockT,
-    C: Send + Sync + 'static + HeaderBackend<B>,
     FL: Send + Sync + FileStorage,
-    FS: Send + Sync + ForestStorage<LookupKey = [u8; 32], Value = Metadata>,
 {
-    async fn send_file(&self, at: Option<<B as BlockT>::Hash>, file_path: String, location: String, owner: AccountId32) -> RpcResult<()> {
-        let _at = at.unwrap_or_else(|| self.client.info().best_hash);
+    async fn upload_file(&self, file_path: String, location: String, owner: AccountId32) -> RpcResult<()> {
 
         let mut file = File::open(PathBuf::from(file_path.clone())).map_err(into_rpc_error)?;
         let mut file_chunks = Vec::new();
@@ -99,24 +101,31 @@ where
             }
         }
 
+        // let mut memdb = MemoryDB::<FL>::default();
+        // let root = Default::default();
+        // {
+        //     let mut t = TrieDBMutBuilder::<FL>::new(&mut memdb, &mut root).build();
+        //     for (chunk_id, chunk) in file_chunks.iter().enumerate() {
+        //         let chunk_id = chunk_id as u64;
+        //         t.insert(chunk_id, chunk).expect("error");
+        //     }
+        // }
+
         let mut file_storage_lock = self.file_storage.write().await;
-        let mut forest_storage_lock = self.forest_storage.write().await;
 
         let fs_metadata = file.metadata().map_err(into_rpc_error)?;
         let file_size = fs_metadata.len();
+
         let file_metadata = Metadata {
             size: file_size,
             // TODO(Arthur/Alexandru): Fingerprint is a missing piece right now.
             // We will get it from `FileData`.
             fingerprint: H256::default(),
             owner: owner.to_string(),
-            location: location.into(),
+            location: location.clone().into(),
         };
 
         let file_key = file_metadata.key();
-        let _ = forest_storage_lock
-        .insert_file_key(&file_key.into(), &file_metadata)
-        .map_err(into_rpc_error)?;
 
         for (chunk_id, chunk) in file_chunks.iter().enumerate() { 
             let chunk_id = chunk_id as u64;
