+++ node/src/rpc.rs
@@ -7,6 +7,8 @@
 
 use std::sync::Arc;
 
+use crate::{services::StorageHubHandlerConfig, types::StorageHubBackend};
+
 use sc_consensus_manual_seal::{
     rpc::{ManualSeal, ManualSealApiServer},
     EngineCommand,
@@ -16,14 +18,12 @@ use storage_hub_runtime::{opaque::Block, AccountId, Balance, Nonce};
 
 pub use sc_rpc::DenyUnsafe;
 use sc_transaction_pool_api::TransactionPool;
-use shc_rpc::FileSystemRpc;
+use shc_rpc::{FileSystemRpc};
 use shc_rpc::FileSystemApiServer;
 
 use sp_api::ProvideRuntimeApi;
 use sp_block_builder::BlockBuilder;
 use sp_blockchain::{Error as BlockChainError, HeaderBackend, HeaderMetadata};
-
-use crate::service::StorageHubBackend;
 
 /// A type representing all RPC extensions.
 pub type RpcExtension = jsonrpsee::RpcModule<()>;
@@ -69,11 +69,11 @@ where
         deny_unsafe,
     } = deps;
 
-    io.merge(System::new(client.clone(), pool, deny_unsafe).into_rpc())?;
+    io.merge(System::new(client.clone(), pool.clone(), deny_unsafe).into_rpc())?;
     io.merge(TransactionPayment::new(client.clone()).into_rpc())?;
 
     if let Some(storage_hub_backend) = storage_hub_backend {
-        io.merge(FileSystemRpc::new(client, storage_hub_backend.file_storage, storage_hub_backend.forest_storage).into_rpc())?;
+        io.merge(FileSystemRpc::new(storage_hub_backend.file_storage).into_rpc())?;
     }
 
     if let Some(command_sink) = command_sink {
