+++ node/src/types.rs
@@ -1,0 +1,19 @@
+use file_manager::in_memory::InMemoryFileStorage;
+use forest_manager::in_memory::InMemoryForestStorage;
+use reference_trie::RefHasher;
+use sp_trie::LayoutV1;
+use tokio::sync::RwLock;
+use std::{sync::Arc};
+
+use crate::services::StorageHubHandlerConfig;
+
+#[derive(Clone)]
+pub struct StorageHubBackend {
+    pub file_storage: Arc<RwLock<InMemoryFileStorage<LayoutV1<RefHasher>>>>,
+    pub forest_storage: Arc<RwLock<InMemoryForestStorage<LayoutV1<RefHasher>>>>
+}
+
+impl StorageHubHandlerConfig for StorageHubBackend {
+    type FileStorage = InMemoryFileStorage<LayoutV1<RefHasher>>;
+    type ForestStorage = InMemoryForestStorage<LayoutV1<RefHasher>>;
+}
